requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
authlib
sqlalchemy
databases
starlette
python-dotenv